pyproject.toml
[project]
name = "backup-google-sheets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.115.11",
    "google-api-python-client>=2.165.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "jinja2>=3.1.6",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
    "requests>=2.31.0",
    "pandas>=2.1.1",
    "openpyxl>=3.1.2",
    "sqlalchemy>=2.0.39",
    "alembic>=1.15.1",
]
